requirements.txt
streamlit
python-dotenv
openai
transformers
torch
requests
rarfile
azure-cognitiveservices-speech
